+++ doc/source/history.rst
@@ -6,7 +6,7 @@ Release history
 0.8.0 (current development version)
 ===================================
 
-Under construction.
+* FIXED: updated the CUDA backend for the change ``async`` -> ``async_`` in the new versions of PyCUDA. Bumped PyOpenCL and PyCUDA versions to 2018.1.1.
 
 
 0.7.2 (16 Sep 2018)